pyproject.toml
[project]
name = "noti"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-celery-results>=2.6.0",
    "django-filter>=24.4",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.2.10",
    "pydantic>=2.12.1",
    "python-dotenv>=1.1.1",
    "redis>=5.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
]
